+++ lib/aws/rails/sqs_active_job/configuration.rb
@@ -80,7 +80,7 @@ module Aws
         end
 
         def client
-          @client ||= Aws::SQS::Client.new
+          @client ||= Aws::SQS::Client.new(user_agent_suffix: user_agent)
         end
 
         # Return the queue_url for a given job_queue name
@@ -141,6 +141,10 @@ module Aws
         # @return [String] Configuration path found in environment or YAML file.
         def config_file_path(options)
           options[:config_file] || ENV["AWS_SQS_ACTIVE_JOB_CONFIG_FILE"]
+        end
+
+        def user_agent
+          "ft/aws-sdk-rails-activejob/#{Aws::Rails::VERSION}"
         end
       end
     end